Move coredata run to new runners package.
